requirements.txt
streamlit
transformers
torch
scipy
pandas
numpy
requests
beautifulsoup4
google-generativeai
plotly
